feat: implement basic socket syscalls
Add initial support for SYS_SOCKET, SYS_CONNECT, SYS_SEND, and
SYS_RECV using lwIP. This includes a basic socket structure and
minimal TCP callback handling. Also includes minor cleanup of
unused variables and warnings in drivers and userspace.

+++ userspace/compositor.c
@@ -413,7 +413,7 @@ static void handle_mouse_event(int x, int y, uint32_t buttons) {
         if (drag_window) {
             // Snapping
             if (mouse_x <= 10) snap_window(drag_window, 1); // Left
-            else if (mouse_x >= fb_info.width - 10) snap_window(drag_window, 2); // Right
+            else if (mouse_x >= (int)fb_info.width - 10) snap_window(drag_window, 2); // Right
             else if (mouse_y <= 10) snap_window(drag_window, 3); // Maximize
         }
         drag_window = NULL;
